rebase

+++ code/modules/hydroponics/grown/kudzu.dm
@@ -6,7 +6,7 @@
 	icon_state = "seed-kudzu"
 	species = "kudzu"
 	plantname = "Kudzu"
-	product = /obj/item/weapon/reagent_containers/food/snacks/grown/kudzupod
+	product = /obj/item/reagent_containers/food/snacks/grown/kudzupod
 	genes = list(/datum/plant_gene/trait/repeated_harvest, /datum/plant_gene/trait/plant_type/weed_hardy)
 	lifespan = 20
 	endurance = 10
@@ -29,15 +29,23 @@
 /obj/item/seeds/kudzu/proc/plant(mob/user)
 	if(isspaceturf(user.loc))
 		return
-	var/turf/T = get_turf(src)
-	message_admins("Kudzu planted by [key_name_admin(user)](<A HREF='?_src_=holder;adminmoreinfo=\ref[user]'>?</A>) (<A HREF='?_src_=holder;adminplayerobservefollow=\ref[user]'>FLW</A>) at ([T.x],[T.y],[T.z] - <A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[T.x];Y=[T.y];Z=[T.z]'>(JMP)</a>)",0,1)
-	investigate_log("was planted by [key_name(user)] at ([T.x],[T.y],[T.z])","kudzu")
-	new /obj/effect/spacevine_controller(user.loc, mutations, potency, production)
+	if(!isturf(user.loc))
+		to_chat(user, "<span class='warning'>You need more space to plant [src].</span>")
+		return FALSE
+	if(locate(/obj/structure/spacevine) in user.loc)
+		to_chat(user, "<span class='warning'>There is too much kudzu here to plant [src].</span>")
+		return FALSE
+	to_chat(user, "<span class='notice'>You plant [src].</span>")
+	message_admins("Kudzu planted by [ADMIN_LOOKUPFLW(user)] at [ADMIN_COORDJMP(user)]",0,1)
+	investigate_log("was planted by [key_name(user)] at [COORD(user)]", INVESTIGATE_BOTANY)
+	new /datum/spacevine_controller(get_turf(user), mutations, potency, production)
 	qdel(src)
 
 /obj/item/seeds/kudzu/attack_self(mob/user)
-	plant(user)
-	user << "<span class='notice'>You plant the kudzu. You monster.</span>"
+	user.visible_message("<span class='danger'>[user] begins throwing seeds on the ground...</span>")
+	if(do_after(user, 50, needhand = TRUE, target = user.drop_location(), progress = TRUE))
+		plant(user)
+		to_chat(user, "<span class='notice'>You plant the kudzu. You monster.</span>")
 
 /obj/item/seeds/kudzu/get_analyzer_text()
 	var/text = ..()
@@ -87,10 +95,11 @@
 		adjust_potency(rand(15, -5))
 
 
-/obj/item/weapon/reagent_containers/food/snacks/grown/kudzupod
+/obj/item/reagent_containers/food/snacks/grown/kudzupod
 	seed = /obj/item/seeds/kudzu
 	name = "kudzu pod"
 	desc = "<I>Pueraria Virallis</I>: An invasive species with vines that rapidly creep and wrap around whatever they contact."
 	icon_state = "kudzupod"
 	filling_color = "#6B8E23"
 	bitesize_mod = 2
+	foodtype = VEGETABLES | GROSS
